Removed unused immersive code path
